backend/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2  # Version compatible
Automat==22.10.0
channels==4.0.0
channels-redis==4.1.0
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
daphne==4.0.0
dj-database-url==1.2.0
Django==4.2.7
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.1
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
Pillow==10.2.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyHamcrest==2.0.4
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2024.1
redis==5.0.1
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
twisted==24.3.0
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.5.0
zope.interface==6.1